checkbutton to css
